move compo test to node_wasm_test.go
